# public/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TPLK004 - Sistem Absensi & Manajemen Akademik API",
    "description": "API Documentation untuk Sistem Absensi QR Code dengan Gamifikasi, Manajemen Tugas, dan Analytics. Sistem ini menyediakan endpoint untuk autentikasi, manajemen absensi, tugas, gamifikasi, dan berbagai fitur akademik lainnya.",
    "version": "1.0.0",
    "contact": {
      "name": "TPLK004 Support Team",
      "email": "admin@unpam.ac.id"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    },
    {
      "url": "https://your-production-url.com",
      "description": "Production Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter token in format: Bearer {token}"
      },
      "sanctum": {
        "type": "apiKey",
        "in": "cookie",
        "name": "laravel_session",
        "description": "Laravel Sanctum session authentication"
      }
    },
...
  },
  "paths": {
...
    "/api/attendance/sessions": {
      "get": {
        "tags": ["Attendance"],
        "summary": "Get all attendance sessions",
        "description": "Mendapatkan daftar sesi absensi dengan filter dan pagination",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "course_id",
            "in": "query",
            "description": "Filter by course ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": ["active", "closed", "scheduled"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sessions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "title": {
                            "type": "string"
                          },
                          "meeting_number": {
                            "type": "integer"
                          },
                          "course_name": {
                            "type": "string"
                          },
                          "is_active": {
                            "type": "boolean"
                          },
                          "start_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "end_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/attendance/scan": {
      "post": {
        "tags": ["Attendance"],
        "summary": "Scan QR code untuk absensi",
        "description": "Submit QR code token untuk mencatat kehadiran",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token"],
                "properties": {
                  "token": {
                    "type": "string",
                    "example": "ABC123XYZ"
                  },
                  "latitude": {
                    "type": "number",
                    "format": "float",
                    "example": -6.2088
                  },
                  "longitude": {
                    "type": "number",
                    "format": "float",
                    "example": 106.8456
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attendance recorded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Absensi berhasil dicatat"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "attendance_id": {
                          "type": "integer"
                        },
                        "points_earned": {
                          "type": "integer"
                        },
                        "badges_unlocked": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired token"
          },
          "409": {
            "description": "Already attended"
          }
        }
      }
    },
...
  }
}